fix: lint エラーを修正し spec ファイルの除外設定を追加
- .oxlintrc.json: *.spec.ts を max-lines-per-function/max-lines 除外対象に追加
- .oxlintrc.json: mcp/minecraft/mcp-tools.ts を MCP override に追加
- src/ltm/ltm-storage.ts: インラインコメント→JSDoc、非null アサーション除去
- src/mcp/core-server.ts: async+throw→Promise.reject()、不要な async 除去
- src/opencode/session-adapter.ts: ストリームヘルパーを stream-helpers.ts に分離
- spec/mcp/tools/mc-memory.spec.ts: no-immediate-mutation 修正
