+++ src/opencode/session-adapter.ts
@@ -1,17 +1,11 @@
 import {
 	createOpencode,
 	type Event,
-	type EventMessageUpdated,
-	type EventSessionCompacted,
-	type EventSessionError,
-	type EventSessionIdle,
 	type McpLocalConfig,
 	type McpRemoteConfig,
 	type OpencodeClient,
-	type Part,
 } from "@opencode-ai/sdk/v2";
 
-import { withTimeout } from "../core/functions.ts";
 import type {
 	OpencodePromptParams,
 	OpencodeSessionEvent,
@@ -19,6 +13,15 @@ import type {
 	PromptResult,
 	TokenUsage,
 } from "../core/types.ts";
+import {
+	abortSession,
+	classifyEvent,
+	extractText,
+	extractTokens,
+	nextStreamEvent,
+	returnStreamOnce,
+	sumTokens,
+} from "./stream-helpers.ts";
 
 export interface OpencodeSessionAdapterConfig {
 	port: number;
@@ -154,155 +157,4 @@ export class OpencodeSessionAdapter implements OpencodeSessionPort {
 		this.closeServer = result.server.close;
 		return this.client;
 	}
-}
-type AbortableAsyncStream<T> = AsyncIterator<T> & {
-	return?: (value?: unknown) => Promise<IteratorResult<T>>;
-};
-const STREAM_RETURNED = Symbol("streamReturned"),
-	STREAM_RETURN_PROMISE = Symbol("streamReturnPromise");
-type StreamReadResult = { type: "event"; value: unknown } | { type: "done" } | { type: "aborted" };
-
-/** signal なしの stream.next() に適用するタイムアウト（5分） */
-const STREAM_NEXT_TIMEOUT_MS = 5 * 60 * 1000;
-
-async function nextStreamEvent(
-	stream: AbortableAsyncStream<unknown>,
-	signal: AbortSignal | undefined,
-	onAbort: () => Promise<void>,
-): Promise<StreamReadResult> {
-	if (!signal) {
-		const result = await withTimeout(
-			stream.next(),
-			STREAM_NEXT_TIMEOUT_MS,
-			"stream.next() timed out after 5 minutes",
-		);
-		return result.done ? { type: "done" } : { type: "event", value: result.value };
-	}
-	return waitForNextStreamEvent(stream, signal, onAbort);
-}
-function waitForNextStreamEvent(
-	stream: AbortableAsyncStream<unknown>,
-	signal: AbortSignal,
-	onAbort: () => Promise<void>,
-): Promise<StreamReadResult> {
-	return new Promise<StreamReadResult>((resolve, reject) => {
-		let settled = false;
-		const finish = (complete: () => void) => {
-			if (settled) return;
-			settled = true;
-			signal.removeEventListener("abort", handleAbort);
-			complete();
-		};
-		const handleAbort = () => {
-			finish(() => {
-				void onAbort();
-				void returnStreamOnce(stream).catch(() => {});
-				resolve({ type: "aborted" });
-			});
-		};
-		signal.addEventListener("abort", handleAbort, { once: true });
-		if (signal.aborted) {
-			handleAbort();
-			return;
-		}
-		void (async () => {
-			try {
-				const result = await stream.next();
-				finish(() =>
-					resolve(result.done ? { type: "done" } : { type: "event", value: result.value }),
-				);
-			} catch (error) {
-				finish(() => reject(error));
-			}
-		})();
-	});
-}
-function returnStreamOnce(stream: AbortableAsyncStream<unknown>): Promise<void> {
-	const managed = stream as AbortableAsyncStream<unknown> & {
-		[STREAM_RETURNED]?: boolean;
-		[STREAM_RETURN_PROMISE]?: Promise<unknown>;
-	};
-	if (managed[STREAM_RETURN_PROMISE]) return managed[STREAM_RETURN_PROMISE] as Promise<void>;
-	if (managed[STREAM_RETURNED]) return Promise.resolve();
-	managed[STREAM_RETURNED] = true;
-	managed[STREAM_RETURN_PROMISE] = managed.return ? managed.return() : Promise.resolve();
-	return managed[STREAM_RETURN_PROMISE] as Promise<void>;
-}
-async function abortSession(oc: OpencodeClient, sessionId: string): Promise<void> {
-	try {
-		const result = await oc.session.abort({ sessionID: sessionId });
-		if (result.error) {
-		}
-	} catch {
-		// 停止経路ではベストエフォート。unhandled rejection にはしない。
-	}
-}
-function classifyEvent(
-	typed: Event,
-	sessionId: string,
-	tokensByMessage: Map<string, TokenUsage>,
-): OpencodeSessionEvent | null {
-	if (typed.type === "message.updated") {
-		accumulateTokens(typed as EventMessageUpdated, sessionId, tokensByMessage);
-		return null;
-	}
-	if (typed.type === "session.idle") {
-		const idle = typed as EventSessionIdle;
-		if (idle.properties.sessionID === sessionId) {
-			return { type: "idle", tokens: sumTokens(tokensByMessage) };
-		}
-	}
-	if (typed.type === "session.compacted") {
-		const compacted = typed as EventSessionCompacted;
-		if (compacted.properties.sessionID === sessionId) return { type: "compacted" };
-	}
-	if (typed.type === "session.error") {
-		const err = typed as EventSessionError;
-		if (err.properties.sessionID === sessionId) {
-			return { type: "error", message: JSON.stringify(err.properties) };
-		}
-	}
-	return null;
-}
-function accumulateTokens(
-	typed: EventMessageUpdated,
-	sessionId: string,
-	tokensByMessage: Map<string, TokenUsage>,
-): void {
-	const info = typed.properties.info;
-	if (info.role === "assistant" && info.sessionID === sessionId) {
-		tokensByMessage.set(info.id, {
-			input: info.tokens.input,
-			output: info.tokens.output,
-			cacheRead: info.tokens.cache?.read ?? 0,
-		});
-	}
-}
-function extractText(parts: Part[]): string {
-	return parts
-		.filter((p): p is Part & { type: "text" } => p.type === "text")
-		.map((p) => p.text)
-		.join("");
-}
-function extractTokens(info: {
-	tokens?: { input: number; output: number; cache?: { read: number } };
-}): TokenUsage | undefined {
-	if (!info.tokens) return undefined;
-	return {
-		input: info.tokens.input,
-		output: info.tokens.output,
-		cacheRead: info.tokens.cache?.read ?? 0,
-	};
-}
-function sumTokens(tokensByMessage: Map<string, TokenUsage>): TokenUsage | undefined {
-	if (tokensByMessage.size === 0) return undefined;
-	let cacheRead = 0;
-	let input = 0;
-	let output = 0;
-	for (const t of tokensByMessage.values()) {
-		input += t.input;
-		output += t.output;
-		cacheRead += t.cacheRead;
-	}
-	return { input, output, cacheRead };
 }
